backend/requirements.txt
# ======================
# Core Framework
# ======================
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.3

# ======================
# Database
# ======================
sqlalchemy>=2.0.0
psycopg2-binary==2.9.9

# ======================
# Authentication
# ======================
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# ======================
# Document Processing
# ======================
pandas==2.2.3
openpyxl==3.1.5
pillow==11.0.0
python-dateutil==2.9.0
pdf2image==1.17.0
opencv-python==4.9.0.80
pytesseract==0.3.10

# ======================
# HTTP / External APIs
# ======================
requests==2.31.0

# ======================
# Testing
# ======================
pytest==8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
responses>=0.24.0
httpx==0.27.0

# ======================
# Logging & Monitoring
# ======================
sentry-sdk[fastapi]>=1.40.0
structlog>=24.1.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp>=1.22.0

# ======================
# Code Quality
# ======================
ruff>=0.1.0
black>=23.12.0
mypy>=1.8.0
pre-commit>=3.6.0
slowapi>=0.1.9
